deprecate top-level imports
